requirements.txt
networkx
matplotlib
google-generativeai
python-dotenv
groq
google-genai
seaborn
ollama==0.6.0
langchain==1.0.2 
langchain-core==1.0.1 
langgraph==1.0.1 
langgraph-checkpoint==3.0.0 
langgraph-prebuilt==1.0.1 
langsmith==0.4.38
langchain-classic==1.0.0 
langchain-text-splitters==1.0.0 
langchain_community==0.4
langchain-ollama
langchain-openai
langchain-groq
langchain-google-genai
tqdm==4.67.1
unstructured==0.18.15
langchain-chroma
py2neo
pandas
langchain_neo4j
streamlit
fastapi
scikit-learn